Add files via upload
Code related to my bachelor thesis. 
The increasing use of Wireless Sensor Networks (WSNs) as Internet of Things (IoT) spreads has turned the importance of choosing efficient devices and network protocols even greater. 

In order to facilitate selection of efficient devices, the study of the behavior of the wireless communication devices is needed. Therefore, a remote testbed was required to be developed. The main aim of this project is to represent data related to measurements of the network. Packages with network data are received directly from the nRF Dongles, the devices to study. These dongles then report to the Odroid M1S which sends the information directly to the system’s database. 
The system represents information related to different measurements (e.g. transmission time, current, throughput…) that requires no physical interaction from the user. In order to obtain that result, different Odroids M1S will be in charge of obtaining data from the wireless communication devices that will be stored in the system’s database in different tables depending on the information they provide. The data received by the central server will be represented on a website so future researchers can make their own studies and make decisions according to the results obtained. 
In addition, a second objective was to upload specific binaries in the wireless communication devices, nRF Dongles. For doing so, a Python script that connects to the wireless devices from the Odroids M1S will be in charge of uploading the compiled code.